+++ source/globals.f90
@@ -55,8 +55,8 @@ module globals
   integer :: nextout       !< Number of next output
 
   ! Generic globals
-  integer :: start_mark     !< Timing mark (start of simulation)
-  integer :: it_mark        !< Timing mark (start of iteration)
+  integer(8) :: start_mark  !< Timing mark (start of simulation)
+  integer(8) :: it_mark     !< Timing mark (start of iteration)
   integer :: logu           !< Logfile unit number
   character(80) :: logfile  !< Logfile name
   character(15) :: host     !< The host on which the code runs
